# api-docs.yml
openapi: 3.1.0
info:
  title: Weather Metrics API
  description: API for ingesting and querying weather sensor metrics
  version: "1.0"
servers:
- url: http://localhost:808
  description: Generated server url
paths:
  /v1/weather/metrics:
    post:
      tags:
      - weather-metrics-controller
      summary: Query aggregated sensor metrics
      description: Query aggregated sensor metrics over a date range.
      operationId: getMetrics
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MetricsRequest"
        required: true
      responses:
        "200":
          description: Metrics returned successfully
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MetricsResponse"
        "400":
          description: Invalid request
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MetricsResponse"
  /v1/sensor/ingest-readings/{sensorId}:
    post:
      tags:
      - sensor-controller
      summary: Ingest sensor metrics
      description: API to ingest sensor metrics
      operationId: ingestReadings
      parameters:
      - name: sensorId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SensorRequest"
        required: true
      responses:
        "200":
          description: Readings ingested successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/IngestionResponse"
        "400":
          description: Invalid request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/IngestionResponse"
components:
  schemas:
    MetricsRequest:
      type: object
      properties:
        sensorIds:
          type: array
          items:
            type: integer
            format: int64
          minItems: 1
        metricsTypes:
          type: array
          items:
            type: string
            enum:
            - TEMPERATURE
            - HUMIDITY
            - WIND_SPEED
        statisticType:
          type: string
          enum:
          - MIN
          - MAX
          - SUM
          - AVG
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
      required:
      - sensorIds
    MetricsResponse:
      type: object
      properties:
        sensorId:
          type: integer
          format: int64
        metricsType:
          type: string
          enum:
          - TEMPERATURE
          - HUMIDITY
          - WIND_SPEED
        metricsValue:
          type: number
          format: double
    MetricReading:
      type: object
      properties:
        metricType:
          type: string
          enum:
          - TEMPERATURE
          - HUMIDITY
          - WIND_SPEED
        metricValue:
          type: number
          format: double
      required:
      - metricType
      - metricValue
    SensorRequest:
      type: object
      properties:
        metricReadings:
          type: array
          items:
            $ref: "#/components/schemas/MetricReading"
          minItems: 1
      required:
      - metricReadings
    IngestionResponse:
      type: object
      properties:
        sensorId:
          type: integer
          format: int64
        sensorResponse:
          type: array
          items:
            $ref: "#/components/schemas/SensorResponse"
    SensorResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        metricType:
          type: string
          enum:
          - TEMPERATURE
          - HUMIDITY
          - WIND_SPEED
        metricsValue:
          type: number
          format: double
        timestamp:
          type: string
          format: date-time